Fixed makefile + added layers +linux support+stuff

+++ inc/EGUI/EGUI_Vector.hpp
@@ -3,13 +3,13 @@
 namespace egui{
 	class Vector2D{
 	public:
-		constexpr double X() const { return _x; };
-		constexpr double Y() const { return _y; };
+		constexpr float X() const { return _x; };
+		constexpr float Y() const { return _y; };
 
-		void setX(double X){ _x = X; };
-		void setY(double Y){ _y = Y; };
+		void setX(float X){ _x = X; };
+		void setY(float Y){ _y = Y; };
 
-		Vector2D(const double X, const double Y) : _x(X), _y(Y) {};
+		Vector2D(const float X, const float Y) : _x(X), _y(Y) {};
 		~Vector2D(){};
 
 		constexpr Vector2D operator+(const Vector2D& v) const {
@@ -21,16 +21,16 @@ namespace egui{
         constexpr Vector2D operator*(const Vector2D& v) const {
             return { _x * v._x, _y * v._y };
         }
-        constexpr Vector2D operator*(double s) const {
+        constexpr Vector2D operator*(float s) const {
             return { _x * s, _y * s };
         }
         constexpr Vector2D operator/(const Vector2D& v) const {
             return { _x / v._x, _y / v._y };
         }
-        constexpr Vector2D operator/(double s) const {
+        constexpr Vector2D operator/(float s) const {
             return { _x / s, _y / s };
         }
 	private:
-		double _x{0}, _y{0};
+		float _x{0}, _y{0};
 	};
 }
